+++ pennylane_lightning/core/_serialize.py
@@ -343,14 +343,17 @@ class QuantumScriptSerializer:
             ]
         )
 
-    def _pauli_sentence(self, observable, wires_map: dict = None):
+    def _pauli_sentence(self, observable, wires_map: dict = None, direct_return=False):
         """Serialize a :class:`pennylane.pauli.PauliSentence` into a Hamiltonian."""
         # Trivial Pauli sentences' items is empty, cannot unpack
-        if not observable:
+        if not observable and not direct_return:
             return self.hamiltonian_obs(np.array([0.0]).astype(self.rtype), [self._ob(Identity(0))])
         pwords, coeffs = zip(*observable.items())
         terms = [self._pauli_word(pw, wires_map) for pw in pwords]
         coeffs = np.array(coeffs).astype(self.rtype)
+
+        if direct_return:
+            return coeffs, terms
 
         if self.split_obs:
             c, o = self._chunk_ham_terms(coeffs, terms, self.split_obs)