+++ pennylane_lightning/core/src/bindings/Bindings.hpp
@@ -352,6 +352,8 @@ void registerBackendAgnosticObservables(py::module_ &m) {
                 return NamedObsT(name, wires);
             }))
         .def("__repr__", &NamedObsT::getObsName)
+        .def("get_base_ob_name", &NamedObsT::getObsNameBase,
+             "Get base name of observables")
         .def("get_wires", &NamedObsT::getWires, "Get wires of observables")
         .def(
             "__eq__",